Support Ray RL lib submissions
Added the possibility to save the environment characteristics (more
specifically Observation and Action spaces) that usually require you to
open a connection with the engine.

Loading this info from file allows to avoid booting up the engine at
initialization of the Ray Algo and one just needs to provide the filepath
as done with config files with Stable Baselines.

When using the file, environment check is automatically deactivated.